pyproject.toml
[tool.poetry]
name = "dodo__stop-sales"
version = "0.1.0"
description = ""
authors = ["usbtypec1 <eldos.baktybekov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"
coverage = "^6.4.1"
pika = "^1.3.0"
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
dodo-lib-usbtypec1 = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "src",
]